Experiments/requirements.txt
appnope==0.1.0
attrdict==2.0.1
backcall==0.1.0
bleach==2.1.3
certifi==2018.4.16
chardet==3.0.4
cycler==0.10.0
decorator==4.3.0
docopt==0.6.2
entrypoints==0.2.3
fire==0.1.3
html5lib==1.0.1
idna==2.7
imageio==2.5.0
inferno==0.1.29
ipdb==0.12
ipykernel==4.8.2
ipython==6.4.0
ipython-genutils==0.2.0
ipywidgets==7.2.1
jedi==0.12.0
Jinja2==2.10
jplephem==2.8
jsonpickle==0.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
kiwisolver==1.0.1
labelme==3.16.2
MarkupSafe==1.0
matplotlib==2.2.2
mistune==0.8.3
munch==2.3.2
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.3
notebook==5.5.0
numpy==1.17.2
opencv-python==3.4.1.15
pandas==0.25.1
pandocfilters==1.4.2
parso==0.2.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==6.1.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
py-cpuinfo==4.0.0
pygame==1.9.4
Pygments==2.2.0
pyparsing==2.2.0
PyQt5==5.13.0
PyQt5-sip==4.19.18
python-dateutil==2.7.3
pytz==2018.4
PyWavelets==1.0.3
PyYAML==5.1
pyzmq==17.0.0
qtconsole==4.3.1
QtPy==1.9.0
requests==2.19.0
sacred==0.7.4
scikit-image==0.15.0
scikit-learn==0.19.1
scipy==1.1.0
seaborn==0.9.0
selenium==3.13.0
Send2Trash==1.5.0
setproctitle==1.1.10
sgp4==1.4
Shapely==1.6.4.post2
simplegeneric==0.8.1
six==1.12.0
sklearn==0.0
skyfield==1.9
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
torch==1.2.0
torchfile==0.1.0
torchsummary==1.5.1
torchvision==0.4.0
tornado==5.0.2
tqdm==4.26.0
traitlets==4.3.2
ujson==1.35
urllib3==1.23
visdom==0.1.8.3
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.48.0
widgetsnbextension==3.2.1
wrapt==1.10.11
